# webapp/contracts/openapi/bundled.yaml
openapi: 3.1.0
info:
  title: BizSim Edu API
  description: 商域 - 商业模拟教育平台后端 API
  version: 1.0.0
paths:
...
  /api/v1/teaching-groups:
    post:
      tags:
      - 体验营
      summary: Create Teaching Group
      operationId: create_teaching_group_api_v1_teaching_groups_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeachingGroupCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TeachingGroupOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/teaching-groups/mine:
    get:
      tags:
      - 体验营
      summary: List My Teaching Groups
      operationId: list_my_teaching_groups_api_v1_teaching_groups_mine_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_List_TeachingGroupOut__'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/teaching-groups/joined:
    get:
      tags:
      - 体验营
      summary: List Joined Groups
      operationId: list_joined_groups_api_v1_teaching_groups_joined_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_List_TeachingGroupOut__'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/teaching-groups/join:
    post:
      tags:
      - 体验营
      summary: Join Teaching Group
      operationId: join_teaching_group_api_v1_teaching_groups_join_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeachingGroupJoin'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TeachingGroupOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/teaching-groups/{group_id}:
    get:
      tags:
      - 体验营
      summary: Get Teaching Group
      operationId: get_teaching_group_api_v1_teaching_groups__group_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer
          title: Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TeachingGroupDetail_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - 体验营
      summary: Update Teaching Group
      operationId: update_teaching_group_api_v1_teaching_groups__group_id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer
          title: Group Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeachingGroupUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TeachingGroupOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/teaching-groups/{group_id}/events:
    get:
      tags:
      - 体验营
      summary: List Group Events
      operationId: list_group_events_api_v1_teaching_groups__group_id__events_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer
          title: Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_List_CompetitionEventOut__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/teaching-groups/{group_id}/announcements:
    get:
      tags:
      - 体验营
      summary: List Announcements
      operationId: list_announcements_api_v1_teaching_groups__group_id__announcements_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer
          title: Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_List_AnnouncementOut__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - 体验营
      summary: Create Announcement
      operationId: create_announcement_api_v1_teaching_groups__group_id__announcements_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer
          title: Group Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnouncementCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AnnouncementOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/teaching-groups/{group_id}/announcements/{announcement_id}/pin:
    patch:
      tags:
      - 体验营
      summary: Pin Announcement
      operationId: pin_announcement_api_v1_teaching_groups__group_id__announcements__announcement_id__pin_patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer
          title: Group Id
      - name: announcement_id
        in: path
        required: true
        schema:
          type: integer
          title: Announcement Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AnnouncementOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/teaching-groups/{group_id}/announcements/{announcement_id}:
    delete:
      tags:
      - 体验营
      summary: Delete Announcement
      operationId: delete_announcement_api_v1_teaching_groups__group_id__announcements__announcement_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer
          title: Group Id
      - name: announcement_id
        in: path
        required: true
        schema:
          type: integer
          title: Announcement Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/teaching-groups/{group_id}/member-progress:
    get:
      tags:
      - 体验营
      summary: Get Member Progress
      operationId: get_member_progress_api_v1_teaching_groups__group_id__member_progress_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer
          title: Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_MemberProgressList_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/teaching-groups/{group_id}/dashboard:
    get:
      tags:
      - 体验营
      summary: Get Camp Dashboard
      operationId: get_camp_dashboard_api_v1_teaching_groups__group_id__dashboard_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer
          title: Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_CampDashboardOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/teaching-groups/{group_id}/groups:
    get:
      tags:
      - 营团分组
      summary: List Groups
      operationId: list_groups_api_v1_teaching_groups__group_id__groups_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer
          title: Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_List_CampGroupDetail__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - 营团分组
      summary: Create Group
      operationId: create_group_api_v1_teaching_groups__group_id__groups_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer
          title: Group Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampGroupCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_CampGroupOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AnnouncementCreate:
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
          title: Title
        content:
          type: string
          minLength: 1
          title: Content
      type: object
      required:
      - title
      - content
      title: AnnouncementCreate
...
    ApiResponse_List_TeachingGroupOut__:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        code:
          type: integer
          title: Code
          default: 0
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/TeachingGroupOut'
            type: array
          - type: 'null'
          title: Data
      type: object
      title: ApiResponse[List[TeachingGroupOut]]
...
    ApiResponse_TeachingGroupOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        code:
          type: integer
          title: Code
          default: 0
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/TeachingGroupOut'
          - type: 'null'
      type: object
      title: ApiResponse[TeachingGroupOut]
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TeachingGroupCreate:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        camp_start_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Camp Start At
        camp_end_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Camp End At
      type: object
      required:
      - name
      title: TeachingGroupCreate
...
    TeachingGroupJoin:
      properties:
        invite_code:
          type: string
          maxLength: 6
          minLength: 6
          title: Invite Code
      type: object
      required:
      - invite_code
      title: TeachingGroupJoin
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login